+++ http4k-core/src/test/kotlin/org/http4k/events/MyEvent.kt
@@ -1,0 +1,5 @@
+package org.http4k.events
+
+data class MyEvent(val value: String = "hello") : Event {
+    val category = EventCategory("Foo")
+}