+++ http4k-format-gson/src/test/kotlin/org/http4k/format/GsonTest.kt
@@ -45,3 +45,5 @@ class GsonTest : JsonContract<JsonElement>(Gson) {
 }
 
 class GsonGenerateDataClassesTest : GenerateDataClassesContract<JsonElement>(Gson)
+
+class GsonAutoEventsTest : AutoJsonEventsContract(Gson)