+++ http4k-format-jackson/src/test/kotlin/org/http4k/format/JacksonTest.kt
@@ -74,3 +74,4 @@ class JacksonTest : JsonContract<JsonNode>(Jackson) {
 
 class JacksonGenerateDataClassesTest : GenerateDataClassesContract<JsonNode>(Jackson)
 
+class JacksonAutoEventsTest : AutoJsonEventsContract(Jackson)