+++ http4k-testing-approval/src/main/kotlin/org/http4k/testing/TestNamer.kt
@@ -13,5 +13,9 @@ interface TestNamer {
             override fun nameFor(testClass: Class<*>, testMethod: Method): String =
                 testClass.packageName.replace('.', '/') + '/' + testClass.simpleName + "." + testMethod.name
         }
+        val MethodOnly = object : TestNamer {
+            override fun nameFor(testClass: Class<*>, testMethod: Method): String =
+                testClass.packageName.replace('.', '/') + '/' + testMethod.name
+        }
     }
 }